requirements.txt
numpy>=1.19.0
pytesseract>=0.3.8
Pillow>=8.0.0
tensorflow>=2.4.0
scikit-learn>=0.24.0
seaborn>=0.11.0
pandas>=1.2.0
matplotlib>=3.3.0 